+++ app/javascript/vue/tasks/digitize/store/actions/loadDigitalization.js
@@ -25,6 +25,7 @@ export default ({ commit, dispatch, state }, coId) =>
 
       if (coObject.collecting_event_id) {
         promises.push(dispatch(ActionNames.GetCollectionEvent, coObject.collecting_event_id))
+        promises.push(dispatch(ActionNames.LoadGeoreferences, coObject.collecting_event_id))
       }
 
       promises.push(dispatch(ActionNames.GetTypeMaterial, coId))